Move main methods to separate files
